`Base` class

the `Base` class used by other objects; useful for extensions of the package, not for basic interactions with the package

Methods

Public methods

Base$new()

Base$get_name()

Base$get_type()

Base$get_model()

Base$set_model()

Base$get_method()

Base$set_method()

Base$get_pi_method()

Base$set_pi_method()

Base$get_level()

Base$set_level()

Base$get_B()

Base$set_B()

Base$get_nb_hidden()

Base$set_nb_hidden()

Base$get_nodes_sim()

Base$set_nodes_sim()

Base$get_activ()

Base$set_activ()

Base$set_engine()

Base$get_engine()

Base$get_params()

Base$get_seed()

Base$set_seed()

Base$summary()

Base$clone()

Method new()

Create a new object.

Usage

Base$new(
  name = "Base",
  type = "none",
  model = NULL,
  method = NULL,
  X_train = NULL,
  y_train = NULL,
  pi_method = c("none", "splitconformal", "kdesplitconformal", "bootsplitconformal",
    "jackknifeplus", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal",
    "surrjackknifeplus"),
  level = 95,
  B = 100,
  nb_hidden = 0,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  engine = NULL,
  params = NULL,
  seed = 123
)

Arguments

name

name of the class

type

type of supervised learning method implemented

model

fitted model

method

supevised learning method

X_train

training set features

y_train

training set response

pi_method

type of prediction interval in c("splitconformal", "kdesplitconformal", "bootsplitconformal", "jackknifeplus", "kdejackknifeplus", "bootjackknifeplus", "surrsplitconformal", "surrjackknifeplus")

level

an integer; the level of confidence

B

an integer; the number of simulations when level is not NULL

nb_hidden

number of nodes in the hidden layer, for construction of a quasi- randomized network

nodes_sim

type of 'simulations' for hidden nodes, if nb_hidden > 0; takes values in c("sobol", "halton", "unif")

activ

activation function's name for the hidden layer, in the construction of a quasi-randomized network; takes values in c("relu", "sigmoid", "tanh", " leakyrelu", "elu", "linear")

engine

contains fit and predict lower-level methods for the given method; do not modify by hand

params

additional parameters passed to method when calling fit

seed

an integer; reproducibility seed for methods that include randomization

Returns

A new `Base` object.

Method get_name()

Usage

Base$get_name()

Method get_type()

Usage

Base$get_type()

Method get_model()

Usage

Base$get_model()

Method set_model()

Usage

Base$set_model(model)

Method get_method()

Usage

Base$get_method()

Method set_method()

Usage

Base$set_method(method)

Method get_pi_method()

Usage

Base$get_pi_method()

Method set_pi_method()

Usage

Base$set_pi_method(pi_method)

Method get_level()

Usage

Base$get_level()

Method set_level()

Usage

Base$set_level(level)

Method get_B()

Usage

Base$get_B()

Method set_B()

Usage

Base$set_B(B)

Method get_nb_hidden()

Usage

Base$get_nb_hidden()

Method set_nb_hidden()

Usage

Base$set_nb_hidden(nb_hidden)

Method get_nodes_sim()

Usage

Base$get_nodes_sim()

Method set_nodes_sim()

Usage

Base$set_nodes_sim(nodes_sim)

Method get_activ()

Usage

Base$get_activ()

Method set_activ()

Usage

Base$set_activ(activ)

Method set_engine()

Usage

Base$set_engine(engine)

Method get_engine()

Usage

Base$get_engine()

Method get_params()

Usage

Base$get_params()

Method get_seed()

Usage

Base$get_seed()

Method set_seed()

Usage

Base$set_seed(seed)

Method summary()

Usage

Base$summary(
  X,
  show_progress = TRUE,
  class_name = NULL,
  class_index = NULL,
  y = NULL,
  type_ci = c("student", "nonparametric", "bootstrap"),
  cl = NULL
)

Method clone()

The objects of this class are cloneable with this method.

Usage

Base$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.