`Classifier` class

The `Classifier` class contains supervised classification models

Details

This class implements models:

lm

Linear model

bcn

see https://www.researchgate.net/publication/380760578_Boosted_Configuration_neural_Networks_for_supervised_classification

extratrees

Extremely Randomized Trees; see https://link.springer.com/article/10.1007/s10994-006-6226-1

glmnet

Elastic Net Regression; see https://glmnet.stanford.edu/

krr

Kernel Ridge Regression; see for example https://www.jstatsoft.org/article/view/v079i03

(but the implementation is different)
ranger

Random Forest; see https://www.jstatsoft.org/article/view/v077i01

ridge

Ridge regression; see https://arxiv.org/pdf/1509.09169

xgboost

a scalable tree boosting system see https://arxiv.org/abs/1603.02754

rvfl

Random Vector Functional Network, see https://www.researchgate.net/publication/332292006_Online_Bayesian_Quasi-Random_functional_link_networks_application_to_the_optimization_of_black_box_functions

Super class

learningmachine::Base -> Classifier

Public fields

name

name of the class

type

type of supervised learning method implemented

model

fitted model

method

supervised learning method in c('lm', 'ranger', 'extratrees', 'ridge', 'bcn', 'glmnet', 'krr', 'xgboost')

X_train

training set features; do not modify by hand

y_train

training set response; do not modify by hand

pi_method

type of prediction set in c("splitconformal", "kdesplitconformal", "bootsplitconformal", "surrsplitconformal")

level

an integer; the level of confidence (default is 95, for 95 per cent) for prediction sets

type_prediction_set

a string; the type of prediction set (currently, only "score" method)

B

an integer; the number of simulations when level is not NULL

nb_hidden

number of nodes in the hidden layer, for construction of a quasi- randomized network

nodes_sim

type of 'simulations' for hidden nodes, if nb_hidden > 0; takes values in c("sobol", "halton", "unif")

activ

activation function's name for the hidden layer, in the construction of a quasi-randomized network; takes values in c("relu", "sigmoid", "tanh", " leakyrelu", "elu", "linear")

engine

contains fit and predic lower-level methods for the given method; do not modify by hand

params

additional parameters passed to method when calling fit do not modify by hand

seed

an integer; reproducibility seed for methods that include randomization

Methods

Public methods

Classifier$new()

Classifier$get_type_prediction_set()

Classifier$set_type_prediction_set()

Classifier$fit()

Classifier$predict_proba()

Classifier$predict()

Classifier$clone()

Inherited methods

learningmachine::Base$get_B()

learningmachine::Base$get_activ()

learningmachine::Base$get_engine()

learningmachine::Base$get_level()

learningmachine::Base$get_method()

learningmachine::Base$get_model()

learningmachine::Base$get_name()

learningmachine::Base$get_nb_hidden()

learningmachine::Base$get_nodes_sim()

learningmachine::Base$get_params()

learningmachine::Base$get_pi_method()

learningmachine::Base$get_seed()

learningmachine::Base$get_type()

learningmachine::Base$set_B()

learningmachine::Base$set_activ()

learningmachine::Base$set_engine()

learningmachine::Base$set_level()

learningmachine::Base$set_method()

learningmachine::Base$set_model()

learningmachine::Base$set_nb_hidden()

learningmachine::Base$set_nodes_sim()

learningmachine::Base$set_pi_method()

learningmachine::Base$set_seed()

learningmachine::Base$summary()

Method new()

Create a new object.

Usage

Classifier$new(
  name = "Classifier",
  type = "classification",
  model = NULL,
  method = NULL,
  X_train = NULL,
  y_train = NULL,
  pi_method = c("none", "kdesplitconformal", "bootsplitconformal", "surrsplitconformal"),
  level = 95,
  type_prediction_set = c("none", "score"),
  B = 100,
  nb_hidden = 0,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  engine = NULL,
  params = NULL,
  seed = 123
)

Returns

A new `Classifier` object.

Method get_type_prediction_set()

Usage

Classifier$get_type_prediction_set()

Method set_type_prediction_set()

Usage

Classifier$set_type_prediction_set(type_prediction_set)

Method fit()

Fit model to training set

Usage

Classifier$fit(X, y, ...)

Arguments

X

a matrix of covariates (i.e explanatory variables)

y

a vector, the response (i.e variable to be explained)

...

additional parameters to learning algorithm (see vignettes)

Method predict_proba()

Usage

Classifier$predict_proba(X)

Method predict()

Usage

Classifier$predict(X)

Method clone()

The objects of this class are cloneable with this method.

Usage

Classifier$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.